travel-agent-api/pyproject.toml
[tool.poetry]
name = "travel-agent-api"
version = "0.1.0"
description = "API avanzata per il sistema Travel Agent con AI"
authors = ["Federico Battisti <fed.battisti@gmail.com>"]
readme = "README.md"
packages = [{include = "travel_agent_api", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.116.1"
uvicorn = {extras = ["standard"], version = "^0.35.0"}
openai = "^1.97.1"
langchain-openai = "^0.3.28"
langchain-community = "^0.3.27"
langchain-core = "^0.3.0"
pydantic = "^2.11.7"
python-dotenv = "^1.1.1"
requests = "^2.32.4"
langgraph = "^0.6.0"
google-search-results = "^2.4.2"
python-multipart = "^0.0.20"
pypdf2 = "3.0.1"
python-docx = "^1.2.0"
pytesseract = "^0.3.13"
pillow = "^11.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
